feat(Monster): re-work
